admin: factor out broker routes
